+++ ui/src/lib/utils/files.ts
@@ -1,6 +1,6 @@
 import { saveAs } from "file-saver";
 import markdownIt from "markdown-it";
-import jsPDF from "jspdf";
+import { marked } from "marked";
 
 const downloadFile = (fileContent: string, fileName: string, type: string) => {
 	saveAs(
@@ -12,22 +12,19 @@ const downloadFile = (fileContent: string, fileName: string, type: string) => {
 
 const downloadPDFFile = async (fileContent: string, fileName: string) => {
 	// Convert Markdown to HTML using marked
+	const renderer = new marked.Renderer();
+	const content = marked(fileContent, { renderer });
+
 	const md = new markdownIt();
 	const convertHtml = md.render(fileContent);
+	console.log({ content, convertHtml });
 
-	const doc = new jsPDF({
-		unit: "mm",
-		format: "a4",
-		orientation: "portrait",
-	});
+	const a = document.createElement("a");
+	a.href = "data:text/html;charset=utf-8," + encodeURIComponent(content);
+	a.download = fileName;
 
-	doc.html(convertHtml, {
-		async callback(doc) {
-			await doc.save(fileName);
-		},
-		x: 15,
-		y: 15,
-	});
+	document.body.appendChild(a);
+	a.click();
 };
 
 const downloadTextFile = (fileContent: string, fileName: string) => {